token-builder/pyproject.toml
[project]
name = "tb2e-token"
version = "0.1.0"
description = "Torchbearer 2E token image generator using DALL-E 3"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "Pillow>=10.0",
    "rembg[cpu]>=2.0",
]

[project.scripts]
tb2e-token = "tb2e_token.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tb2e_token"]
